fix(release): retire orphaned release-check graph (#18145)
Workflow consolidation removed the release contract steps and helper
graph, but a later merge resolution restored release-check.ts, its
wrapper/root aliases, the regression-matrix validator, and matrix JSON
without their owners. Public commands fail deterministically and app-core
tests cannot import a deleted helper.

Extracted the live macOS stager failure-block invariant
(requiredMacStaplerFailureBlock + containsContiguousBlock) into
scripts/lib/mac-stager-contract.ts so the #17680 gate stays active.
Removed the dead monolith (1553 lines) and all orphaned wrappers.

AI provider/model: zai / glm-5.2
Client / agent tooling: Hermes Agent
Lane: hermes-t3rpz
