+++ package.json
@@ -105,7 +105,6 @@
     "release": "bunx lerna publish from-package --dist-tag latest --force-publish --yes --no-verify-access --no-sort",
     "release:candidate": "node packages/scripts/release-candidate.mjs",
     "release:candidate:verify": "node packages/scripts/release-candidate.mjs verify",
-    "release:check": "node packages/app-core/scripts/run-release-check.mjs",
     "release:next": "bunx lerna publish from-package --dist-tag next --force-publish --yes --no-verify-access --no-sort",
     "release:beta": "bunx lerna publish from-package --dist-tag beta --force-publish --yes --no-verify-access --no-sort",
     "version:beta": "bunx lerna version prerelease --preid beta --force-publish --yes --no-push --no-git-tag-version",
@@ -140,7 +139,6 @@
     "test:console": "node packages/scripts/test-console/server.mjs",
     "test:apple-entitlements": "bun run --cwd packages/app-core test:apple-entitlements",
     "test:cache-stability": "node packages/scripts/run-vitest.mjs run packages/core/src/runtime/__tests__/cache-key-stability.test.ts",
-    "test:regression-matrix:release": "node packages/app-core/scripts/run-eliza-app-core-script.mjs validate-regression-matrix.mjs --workflow release",
     "test:plugins": "bun run build:core && TEST_PACKAGE_FILTER='\\(plugins/' TEST_SCRIPT_FILTER='^test$' node packages/scripts/run-all-tests.mjs --only=test --no-cloud --concurrency=3 --require-work",
     "test:client": "bun run build:core && TEST_SCRIPT_FILTER='^test$' node packages/scripts/run-all-tests.mjs --lane=client --no-cloud --concurrency=3 --require-work",
     "test:core": "node packages/scripts/with-test-runtime.mjs node packages/scripts/run-all-tests.mjs --only=test --no-cloud --filter='@elizaos/core \\(packages/core\\)#test' --require-work",
