--- packages/app-core/scripts/run-release-check.mjs
@@ -1,23 +1,0 @@
-#!/usr/bin/env node
-/**
- * Node wrapper that runs release-check.ts under the tsx loader from the repo
- * root and forwards its exit status.
- */
-import { spawnSync } from "node:child_process";
-import { dirname, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
-
-const scriptsDir = dirname(fileURLToPath(import.meta.url));
-const repoRoot = resolve(scriptsDir, "../../..");
-const result = spawnSync(
-  process.execPath,
-  ["--import", "tsx", resolve(scriptsDir, "release-check.ts")],
-  {
-    cwd: repoRoot,
-    env: process.env,
-    stdio: "inherit",
-  },
-);
-
-if (result.error) throw result.error;
-process.exit(result.status ?? 1);
